feat: implement getQuestionByRole endpoint and update application routing
