+++ src/controller/auth/verify-invitation-token.ts
@@ -1,8 +1,8 @@
 import { and, eq, gt, sql } from "drizzle-orm";
 import type { NextFunction, Request, Response } from "express";
 import { db } from "../../db";
-import { oneTimeTokens } from "../../db/schema/auth";
 import { invitedUsers } from "../../db/schema/application";
+import { oneTimeTokens } from "../../db/schema/auth";
 
 export async function verifyInvitationToken(
 	req: Request,
@@ -41,13 +41,17 @@ export async function verifyInvitationToken(
 			.set({ updatedAt: sql`NOW()` })
 			.where(eq(oneTimeTokens.id, token[0].id));
 
-		const invitedUserResult = await db.select({invitedUsersId: invitedUsers.id}).from(invitedUsers).where(
-			and(
-				eq(invitedUsers.oneTimeTokensId, token[0].id)
-			),
-		);
+		const invitedUserResult = await db
+			.select({ invitedUsersId: invitedUsers.id })
+			.from(invitedUsers)
+			.where(and(eq(invitedUsers.oneTimeTokensId, token[0].id)));
 
-		res.status(200).json({ message: "The token successfully verified", data: {invitedUsersId: invitedUserResult[0].invitedUsersId}  });
+		res
+			.status(200)
+			.json({
+				message: "The token successfully verified",
+				data: { invitedUsersId: invitedUserResult[0].invitedUsersId },
+			});
 		return;
 	} catch (error) {
 		next(error);
