+++ src/routes/application-route.ts
@@ -1,0 +1,6 @@
+import express from "express";
+import { getQuestionByRole } from "../controller/application/get-question-by-role";
+
+export const applicationRoutes = express.Router();
+
+applicationRoutes.get("/get-application-question", getQuestionByRole);
